python/feishu_robot_sg/requirements.txt
Flask==2.0.2
python-dotenv
requests
pycryptodome
tencentcloud-sdk-python-vpc
tencentcloud-sdk-python
gevent